// app/services/enums.ts
enum Condition {
  NEW = "new",
  LIKE_NEW = "like_new",
  USED = "used",
  HEAVILY_USED = "heavily_used",
  BROKEN = "broken"
}

enum Category {
  AIRSOFT_GUN = "airsoft_gun",
  ACCESSORY = "accessory",
  CLOTHING = "clothing",
  GEAR = "gear",
  BATTERY = "battery",
  MAGAZINE = "magazine",
  OPTICS = "optics",
  PATCH = "patch",
  GAS = "gas",
  PART = "part",
  OTHER = "other"
}

export { Condition, Category };